+++ OpenRA.Mods.Common/Traits/World/EditorResourceLayer.cs
@@ -81,14 +81,14 @@ namespace OpenRA.Mods.Common.Traits
 		public event Action<CPos, string> CellChanged;
 
 		ResourceLayerContents IResourceLayer.GetResource(CPos cell) { return Tiles.Contains(cell) ? Tiles[cell] : default; }
-		int IResourceLayer.GetMaxDensity(string resourceType)
+		byte IResourceLayer.GetMaxDensity(string resourceType)
 		{
-			return info.ResourceTypes.TryGetValue(resourceType, out var resourceInfo) ? resourceInfo.MaxDensity : 0;
+			return info.ResourceTypes.TryGetValue(resourceType, out var resourceInfo) ? resourceInfo.MaxDensity : (byte)0;
 		}
 
-		bool IResourceLayer.CanAddResource(string resourceType, CPos cell, int amount) { return CanAddResource(resourceType, cell, amount); }
-		int IResourceLayer.AddResource(string resourceType, CPos cell, int amount) { return AddResource(resourceType, cell, amount); }
-		int IResourceLayer.RemoveResource(string resourceType, CPos cell, int amount) { return RemoveResource(resourceType, cell, amount); }
+		bool IResourceLayer.CanAddResource(string resourceType, CPos cell, byte amount) { return CanAddResource(resourceType, cell, amount); }
+		int IResourceLayer.AddResource(string resourceType, CPos cell, byte amount) { return AddResource(resourceType, cell, amount); }
+		int IResourceLayer.RemoveResource(string resourceType, CPos cell, byte amount) { return RemoveResource(resourceType, cell, amount); }
 		void IResourceLayer.ClearResources(CPos cell) { ClearResources(cell); }
 		bool IResourceLayer.IsVisible(CPos cell) { return Map.Contains(cell); }
 		bool IResourceLayer.IsEmpty => false;
@@ -166,7 +166,7 @@ namespace OpenRA.Mods.Common.Traits
 			}
 		}
 
-		void UpdateNetWorth(string oldResourceType, int oldDensity, string newResourceType, int newDensity)
+		void UpdateNetWorth(string oldResourceType, byte oldDensity, string newResourceType, byte newDensity)
 		{
 			if (oldResourceType != null && oldDensity > 0 && resourceValues.TryGetValue(oldResourceType, out var oldResourceValue))
 				NetWorth -= oldDensity * oldResourceValue;
@@ -197,14 +197,14 @@ namespace OpenRA.Mods.Common.Traits
 		/// <summary>
 		/// Matches the logic in <see cref="ResourceLayer"/> trait.
 		/// </summary>
-		protected virtual int CalculateCellDensity(ResourceLayerContents contents, CPos cell)
+		protected virtual byte CalculateCellDensity(ResourceLayerContents contents, CPos cell)
 		{
 			var resources = Map.Resources;
 			if (contents.Type == null || !info.ResourceTypes.TryGetValue(contents.Type, out var resourceInfo) || resources[cell].Type != resourceInfo.ResourceIndex)
 				return 0;
 
 			if (!info.RecalculateResourceDensity)
-				return contents.Density.Clamp(1, resourceInfo.MaxDensity);
+				return contents.Density.Clamp((byte)1, resourceInfo.MaxDensity);
 
 			// Set density based on the number of neighboring resources.
 			var adjacent = 0;
@@ -219,7 +219,7 @@ namespace OpenRA.Mods.Common.Traits
 			// We need to have at least one resource in the cell.
 			// HACK: we should not be lerping to 9, as maximum adjacent resources is 8.
 			// HACK: it's too disruptive to fix.
-			return Math.Max(int2.Lerp(0, resourceInfo.MaxDensity, adjacent, 9), 1);
+			return (byte)Math.Max(int2.Lerp(0, resourceInfo.MaxDensity, adjacent, 9), 1);
 		}
 
 		protected virtual bool AllowResourceAt(string resourceType, CPos cell)
