+++ OpenRA.Mods.Common/TraitsInterfaces.cs
@@ -808,10 +808,10 @@ namespace OpenRA.Mods.Common.Traits
 	{
 		event Action<CPos, string> CellChanged;
 		ResourceLayerContents GetResource(CPos cell);
-		int GetMaxDensity(string resourceType);
-		bool CanAddResource(string resourceType, CPos cell, int amount = 1);
-		int AddResource(string resourceType, CPos cell, int amount = 1);
-		int RemoveResource(string resourceType, CPos cell, int amount = 1);
+		byte GetMaxDensity(string resourceType);
+		bool CanAddResource(string resourceType, CPos cell, byte amount = 1);
+		int AddResource(string resourceType, CPos cell, byte amount = 1);
+		int RemoveResource(string resourceType, CPos cell, byte amount = 1);
 		void ClearResources(CPos cell);
 
 		bool IsVisible(CPos cell);
